requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.7
django-appconf==1.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
django-googledrive-storage==1.3.4
django-modeltranslation==0.12.2
djangorestframework==3.7.3
generic==0.3.1
google-api-python-client==1.6.5
httplib2==0.10.3
idna==2.6
mysqlclient==1.3.12
oauth2client==4.1.2
olefile==0.44
Pillow==4.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyDrive==1.3.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
